minor tweaks

+++ task1-B1/test.cpp
@@ -3,8 +3,12 @@
 #include "string_functions.hpp"
 
 TEST_CASE("SimpleTests") {
-    REQUIRE(GetStringByPrefixFunction(std::vector<int>({0, 1, 2, 3, 0})) == "aaaab");
-    REQUIRE(GetStringByPrefixFunction(std::vector<int>({0})) == "a");
-    REQUIRE(GetStringByPrefixFunction(std::vector<int>({0, 0, 1, 0, 1, 2, 3, 0})) == "abacabad");
-    REQUIRE(GetStringByPrefixFunction(std::vector<int>({0, 0, 1, 2, 3, 4, 0, 1})) == "abababba");
+    REQUIRE(GetStringByPrefixFunction(
+        std::vector<size_t>({0, 1, 2, 3, 0})) == "aaaab");
+    REQUIRE(GetStringByPrefixFunction(
+        std::vector<size_t>({0})) == "a");
+    REQUIRE(GetStringByPrefixFunction(
+        std::vector<size_t>({0, 0, 1, 0, 1, 2, 3, 0})) == "abacabad");
+    REQUIRE(GetStringByPrefixFunction(
+        std::vector<size_t>({0, 0, 1, 2, 3, 4, 0, 1})) == "abababba");
 }
